pyproject.toml
[project]
name = "borg"
version = "0.1.0"
description = "Project Borg — PostgreSQL-native memory compiler for AI workflows"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.0",
    "fastmcp>=3.1.0,<3.2",
    "uvicorn[standard]>=0.34.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "httpx>=0.28.0",
    "PyJWT[crypto]>=2.10.1",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.7.0",
    "openai>=1.75.0",
    "tiktoken>=0.8.0",
    "python-dotenv>=1.0.1",
    "structlog>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-httpx>=0.35.0",
    "ruff==0.15.6",
    "mypy>=1.14.0",
]
bootstrap = [
    "httpx>=0.28.0",
]
cli = [
    "click>=8.0",
    "jinja2>=3.0",
    "httpx>=0.27",
]
bench = [
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "openai>=1.75.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
borg = "cli.borg_cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/borg", "cli"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "S", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S106", "S108"]
"cli/**/*.py" = ["S108", "S701"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: end-to-end CLI integration tests (may require a running Borg engine for some tests)",
]
